move utils for model
